cloud/sam/scripts/chat_function/requirements.txt
requests
urllib3==1.26
langchain==0.0.234
llama-index==0.7.9
openai==0.27.8
python-dotenv